Introduce the ability to pause and resume events by their hook.
Pausing an event will disable all actions attached to the event's hook. The event itself will remain in place and will run according to its schedule, but all actions attached to its hook will be disabled. This renders the event inoperative but keeps it scheduled so as to remain fully compatible with plugins which automatically schedule events when they're missing.